requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pymupdf==1.24.9
regex==2024.5.15
python-chess==1.999
openai>=1.0.0
anthropic>=0.40.0
PyYAML==6.0.2
pydantic==2.8.2
python-dotenv==1.0.1